refactor: 기존 BlockPort인터페이스가 BlockRepository와 같은 역하을 하고 있어 다시 원래대로 수정
